requirements.txt
docling==2.46.0
fastapi==0.116.1
protobuf==6.32.0
python-dotenv==1.1.1
uvicorn==0.35.0
google-generativeai